Phase 5: Extract Quick Practice tab into src/ui/quick_practice_tab.py (v7.1.8-claude-dev-phase5)
- New src/ui/quick_practice_tab.py (~390 lines):
    render_quick_practice_tab()  — top-level entry point
    _render_materials_loader()   — Load Practice Materials expander
    _render_builtin_materials()  — built-in library browser + enrichment UI
    _render_upload_materials()   — file upload + enrichment + server save
    _render_practice_area()      — material display + mode dispatch
    _render_guided_mode()        — guided list navigation, diagnostics, phrase display
    _render_free_text_mode()     — free-text input
- src/app.py: Tab 1 inline block (~760 lines) replaced with render_quick_practice_tab()
  call; app.py now 872 lines (down from 2808 at start of refactor)
- tests/test_ui.py: 5 new smoke tests for quick_practice_tab exports
- Version bumped to 7.1.8-claude-dev-phase5

73/73 tests pass. App imports cleanly.
