+++ APP_CHANGELOG.md
@@ -8,6 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [7.1.8-claude-dev-phase5] - 2026-04-05
+
+### Changed
+
+- Version bump
+
+
 ## [7.1.7-claude-dev-phase4] - 2026-04-05
 
 ### Changed
