+++ src/app.py
@@ -96,6 +96,7 @@ from ui.practice_tab import (
 from ui.story_tab import render_story_reader
 from ui.statistics_tab import render_statistics_tab
 from ui.history_tab import load_history, save_history, render_history_tab
+from ui.quick_practice_tab import render_quick_practice_tab
 
 # Import API usage logger for cost tracking
 try:
@@ -848,766 +849,7 @@ def main():
 
     # Tab 1: Quick Practice
     if selected_tab_index == 0:
-        st.header("Quick Practice")
-
-        # Language announcement banner
-        # st.info("🎉 **Now supporting 6 languages!** Practice pronunciation in Portuguese, French, Dutch, German, Italian, and Spanish.")
-
-        # Help info for new users
-        current_session = st.session_state.current_sessions[st.session_state.language]
-        if len(current_session["practices"]) == 0:
-            st.info("👋 **New here?** Check the [User Guide](https://github.com/fairflow/miolingo/blob/feature/admin-fusion/docs/app-docs/USER_GUIDE.md) for step-by-step instructions!")
-
-        # Phrase/Word list loading - Built-in Library + User Upload
-        # Keep expander open during interactions
-        if 'qp_materials_expanded' not in st.session_state:
-            st.session_state.qp_materials_expanded = False
-
-        with st.expander("📚 Load Practice Materials", expanded=st.session_state.qp_materials_expanded):
-            # Set to expanded when user opens it
-            st.session_state.qp_materials_expanded = True
-
-            from app_language_materials import (
-                get_available_languages,
-                get_language_structure,
-                get_file_metadata,
-                load_phrase_file,
-                format_category_name,
-                format_language_name
-            )
-
-            # Use radio buttons for material source to preserve state across reruns
-            material_source_names = ["📦 Built-in Library", "📁 Upload File"]
-            material_source_index = st.radio(
-                "Material Source",
-                range(len(material_source_names)),
-                format_func=lambda i: material_source_names[i],
-                key='material_source_tab',
-                horizontal=True,
-                label_visibility='collapsed'
-            )
-
-            # Show content based on selection
-            if material_source_index == 0:
-                # Built-in materials
-                st.write("Browse curated phrase and word lists by language and level.")
-
-                languages = get_available_languages()
-
-                if not languages:
-                    st.warning("No built-in materials found in `language_materials/` directory.")
-                else:
-                    # Use material language from sidebar session state
-                    material_lang = st.session_state.get('material_language', 'fr')
-
-                    # Show which material language is active
-                    st.info(f"📚 Loading materials for: **{format_language_name(material_lang)}** (change in sidebar)")
-
-                    structure = get_language_structure(material_lang)
-
-                    if structure:
-                        # Category selection (phrases vs words, level A-D)
-                        categories = list(structure.keys())
-                        category = st.selectbox(
-                            "Category",
-                            categories,
-                            format_func=format_category_name,
-                            help="Select difficulty level: Beginner (A) → Expert (D)"
-                        )
-
-                        # File selection within category
-                        files = structure[category]
-                        selected_file = st.selectbox(
-                            "File",
-                            files,
-                            help="Select a specific file from this category"
-                        )
-
-                        # Show metadata preview
-                        metadata = get_file_metadata(material_lang, category, selected_file)
-
-                        if metadata and 'line_count' in metadata:
-                            col1, col2, col3 = st.columns(3)
-                            with col1:
-                                st.metric("Items", metadata.get('line_count', 0))
-                            with col2:
-                                st.metric("Translations", "✓" if metadata.get('has_translations') else "✗")
-                            with col3:
-                                st.metric("IPA", "✓" if metadata.get('has_ipa') else "✗")
-
-                            # Preview
-                            if metadata.get('preview'):
-                                with st.expander("Preview first 3 items"):
-                                    for line in metadata['preview']:
-                                        st.text(line)
-
-                            # Material Enrichment UI
-                            missing_translations = not metadata.get('has_translations')
-                            missing_ipa = not metadata.get('has_ipa')
-
-                            if missing_translations or missing_ipa:
-                                st.markdown("---")
-                                st.markdown("**✨ Enrich This Material**")
-
-                                # Enrichment options
-                                enrich_col1, enrich_col2 = st.columns(2)
-                                with enrich_col1:
-                                    add_trans = st.checkbox(
-                                        "Add translations",
-                                        value=missing_translations,
-                                        disabled=not missing_translations,
-                                        key=f"enrich_trans_{selected_file}"
-                                    )
-                                with enrich_col2:
-                                    add_ipa_check = st.checkbox(
-                                        "Add IPA",
-                                        value=missing_ipa,
-                                        disabled=not missing_ipa,
-                                        key=f"enrich_ipa_{selected_file}"
-                                    )
-
-                                if st.button(
-                                    "✨ Enrich Material",
-                                    type="secondary",
-                                    disabled=not (add_trans or add_ipa_check),
-                                    key=f"enrich_btn_{selected_file}"
-                                ):
-                                    with st.spinner("Enriching material... This may take a minute for translations."):
-                                        # Progress bar
-                                        progress_bar = st.progress(0)
-                                        status_text = st.empty()
-
-                                        def progress_callback(current, total, message):
-                                            progress = current / total if total > 0 else 0
-                                            progress_bar.progress(min(progress, 1.0))
-                                            status_text.text(message)
-
-                                        # Perform enrichment
-                                        result = enrich_material_file(
-                                            file_path=metadata['path'],
-                                            lang_code=material_lang,
-                                            add_translations=add_trans,
-                                            add_ipa=add_ipa_check,
-                                            progress_callback=progress_callback
-                                        )
-
-                                        progress_bar.empty()
-                                        status_text.empty()
-
-                                        if result['success']:
-                                            stats = result['stats']
-                                            st.success(f"✅ Material enriched successfully!")
-
-                                            # Show stats
-                                            stat_col1, stat_col2, stat_col3 = st.columns(3)
-                                            with stat_col1:
-                                                st.metric("Lines processed", stats.get('total_lines', 0))
-                                            with stat_col2:
-                                                st.metric("Translations added", stats.get('translations_added', 0))
-                                            with stat_col3:
-                                                st.metric("IPA added", stats.get('ipa_added', 0))
-
-                                            # Show errors if any
-                                            if stats.get('errors'):
-                                                with st.expander(f"⚠️ {len(stats['errors'])} error(s) occurred"):
-                                                    for error in stats['errors'][:10]:  # Show first 10
-                                                        st.text(error)
-
-                                            st.info("📝 Original file backed up to .bak. Click reload to see updated checkmarks.")
-
-                                            # Suggest reloading
-                                            if st.button("🔄 Reload Metadata", key=f"reload_meta_{selected_file}"):
-                                                # Clear all cached data to force reload of file metadata
-                                                st.cache_data.clear()
-                                                # active_tab already preserved by radio key
-                                                st.rerun()
-                                        else:
-                                            st.error(f"❌ Enrichment failed: {result['message']}")
-                                            if result.get('stats', {}).get('errors'):
-                                                with st.expander("Error details"):
-                                                    for error in result['stats']['errors']:
-                                                        st.text(error)
-
-                                st.markdown("---")
-
-                            # Load button
-                            if st.button("📂 Load This File", type="primary", key="load_builtin"):
-                                try:
-                                    phrases = load_phrase_file(str(metadata['path']))
-                                    st.session_state.phrase_list = phrases
-                                    st.session_state.qp_phrase_position = 0
-                                    st.session_state.phrase_selector_widget = 0  # Sync widget state
-                                    st.session_state.state_change_log.append(f"Load builtin: Reset position to 0")
-                                    st.session_state.quick_last_result = None
-                                    st.session_state.material_source = f"{format_language_name(material_lang)} - {format_category_name(category)} - {selected_file}"
-                                    st.session_state.qp_materials_expanded = False  # Close expander
-                                    st.success(f"✓ Loaded {len(phrases)} items - scroll down to practice section")
-                                    # NO rerun - let Streamlit update naturally to preserve tab
-                                except Exception as e:
-                                    st.error(f"Error loading file: {e}")
-                        else:
-                            st.error("Could not read file metadata")
-                    else:
-                        st.info(f"No materials found for {format_language_name(material_lang)}")
-
-            elif material_source_index == 1:
-                # User upload
-                st.write("Upload your own phrase or word list.")
-                st.caption("**Format:** One phrase per line, or `phrase | translation | [ipa]`")
-                st.caption("**Limits:** Max 200 lines, 200 chars per line")
-
-                # File upload size limits
-                MAX_UPLOAD_LINES = 200
-                MAX_LINE_LENGTH = 200
-
-                uploaded_file = st.file_uploader(
-                    "Choose a text file",
-                    type=['txt'],
-                    help="Upload a .txt file with one phrase per line. Empty lines and comments (#) are ignored."
-                )
-
-                if uploaded_file is not None:
-                    try:
-                        # Read content
-                        content = uploaded_file.read().decode('utf-8')
-
-                        # Store in session state (original or enriched version)
-                        upload_key = f"upload_{uploaded_file.name}_{uploaded_file.size}"
-                        if upload_key not in st.session_state:
-                            st.session_state[upload_key] = content
-                        else:
-                            # Use stored version (may be enriched)
-                            content = st.session_state[upload_key]
-
-                        # Validate size limits (parse from current content, not original)
-                        raw_lines = [line.strip() for line in content.split('\n') if line.strip() and not line.strip().startswith('#')]
-
-                        if len(raw_lines) > MAX_UPLOAD_LINES:
-                            st.error(f"❌ File too large: {len(raw_lines)} lines (max {MAX_UPLOAD_LINES})")
-                            st.stop()
-
-                        for i, line in enumerate(raw_lines, 1):
-                            # Check the original phrase part only (before first |)
-                            phrase_part = line.split('|')[0].strip()
-                            if len(phrase_part) > MAX_LINE_LENGTH:
-                                st.error(f"❌ Line {i} too long: {len(phrase_part)} chars (max {MAX_LINE_LENGTH})")
-                                st.stop()
-
-                        # Parse phrases - support both simple and enhanced format
-                        phrases = []
-                        has_translations = False
-                        has_ipa = False
-
-                        for line in raw_lines:
-                            # Skip comments
-                            if line.startswith('#'):
-                                continue
-
-                            if '|' in line:
-                                # Enhanced format with translation
-                                parts = [p.strip() for p in line.split('|')]
-                                phrase_dict = {
-                                    'text': parts[0],
-                                    'translation': parts[1] if len(parts) > 1 and parts[1] else None,
-                                    'ipa': parts[2] if len(parts) > 2 and parts[2] else None
-                                }
-                                if phrase_dict['translation']:
-                                    has_translations = True
-                                if phrase_dict['ipa']:
-                                    has_ipa = True
-                                phrases.append(phrase_dict)
-                            else:
-                                # Simple format - just the text
-                                phrases.append({'text': line, 'translation': None, 'ipa': None})
-
-                        st.success(f"✓ Loaded {len(phrases)} items from upload")
-
-                        # Show metadata and preview
-                        col1, col2, col3 = st.columns(3)
-                        with col1:
-                            st.metric("Phrases", len(phrases))
-                        with col2:
-                            st.metric("Translations", "✓" if has_translations else "✗")
-                        with col3:
-                            st.metric("IPA", "✓" if has_ipa else "✗")
-
-                        # Show saved status
-                        if st.session_state.get(f"{upload_key}_saved"):
-                            st.success("💾 Saved to server - showing saved version")
-
-                        # Preview
-                        with st.expander("📋 Preview", expanded=True):
-                            preview_count = min(5, len(phrases))
-                            for i, p in enumerate(phrases[:preview_count]):
-                                if p.get('translation') or p.get('ipa'):
-                                    # Always show all 3 fields to match file format
-                                    translation = p.get('translation') or ''
-                                    ipa = p.get('ipa') or ''
-                                    st.text(f"{p['text']} | {translation} | {ipa}")
-                                else:
-                                    st.text(p['text'])
-                            if len(phrases) > preview_count:
-                                st.caption(f"...and {len(phrases) - preview_count} more")
-
-                        # Raw content view for debugging
-                        with st.expander("🔍 Raw File Content (first 5 lines)", expanded=False):
-                            st.caption("This shows the actual file content in session state:")
-                            raw_lines = [line for line in content.split('\n')[:5] if line.strip() and not line.strip().startswith('#')]
-                            for line in raw_lines:
-                                st.code(line, language=None)
-
-                        # Enrichment UI for uploaded files
-                        missing_translations = not has_translations
-                        missing_ipa = not has_ipa
-
-                        if missing_translations or missing_ipa:
-                            st.markdown("---")
-                            st.markdown("**✨ Enrich This Material**")
-                            st.caption("💡 Add translations and/or IPA pronunciation to your uploaded file using AI")
-
-                            enrich_col1, enrich_col2 = st.columns(2)
-                            with enrich_col1:
-                                add_trans_upload = st.checkbox(
-                                    "Add translations",
-                                    value=missing_translations,
-                                    disabled=not missing_translations,
-                                    key="enrich_trans_upload"
-                                )
-                            with enrich_col2:
-                                add_ipa_upload = st.checkbox(
-                                    "Add IPA",
-                                    value=missing_ipa,
-                                    disabled=not missing_ipa,
-                                    key="enrich_ipa_upload"
-                                )
-
-                            if st.button("✨ Enrich Now", type="secondary", key="enrich_upload_btn"):
-                                if add_trans_upload or add_ipa_upload:
-                                    with st.spinner("Enriching material... This may take a minute."):
-                                        # Enrich the phrases in memory
-                                        import tempfile
-                                        import os
-                                        from pathlib import Path
-
-                                        # Save to temp file for enrichment
-                                        with tempfile.NamedTemporaryFile(mode='w', suffix='.txt', delete=False) as tf:
-                                            for p in phrases:
-                                                parts = [p['text']]
-                                                if p.get('translation'):
-                                                    parts.append(p['translation'])
-                                                if p.get('ipa'):
-                                                    parts.append(p['ipa'])
-                                                tf.write(' | '.join(parts) + '\n')
-                                            temp_path = Path(tf.name)
-
-                                        try:
-                                            progress_bar = st.progress(0)
-                                            status_text = st.empty()
-
-                                            def progress_callback(current, total, message):
-                                                progress = current / total if total > 0 else 0
-                                                progress_bar.progress(min(progress, 1.0))
-                                                status_text.text(message)
-
-                                            # Enrich the temp file
-                                            # Use material_language (short code like 'pt', 'fr') not language (full name)
-                                            current_lang = st.session_state.get('material_language', 'fr')
-                                            result = enrich_material_file(
-                                                file_path=temp_path,
-                                                lang_code=current_lang,
-                                                add_translations=add_trans_upload,
-                                                add_ipa=add_ipa_upload,
-                                                progress_callback=progress_callback
-                                            )
-
-                                            progress_bar.empty()
-                                            status_text.empty()
-
-                                            if result['success']:
-                                                # Read enriched content and update session state
-                                                with open(temp_path, 'r', encoding='utf-8') as f:
-                                                    enriched_content = f.read()
-
-                                                # Update session state with enriched version
-                                                st.session_state[upload_key] = enriched_content
-
-                                                # Show enrichment stats
-                                                stats = result.get('stats', {})
-                                                success_msg = st.success(f"✅ Enriched: {stats.get('translations_added', 0)} translations, {stats.get('ipa_added', 0)} IPA")
-
-                                                # Show any errors
-                                                errors = stats.get('errors', [])
-                                                if errors:
-                                                    with st.expander(f"⚠️ {len(errors)} errors occurred", expanded=True):
-                                                        for err in errors[:20]:  # Show first 20
-                                                            st.caption(err)
-                                                        if len(errors) > 20:
-                                                            st.caption(f"...and {len(errors) - 20} more")
-
-                                                # Show preview of first enriched lines for debugging
-                                                first_lines = [line for line in enriched_content.split('\n')[:5] if line.strip() and not line.startswith('#')]
-                                                with st.expander("🔍 First enriched lines (actual file content)", expanded=True):
-                                                    st.caption("This shows what was actually written to the file:")
-                                                    for line in first_lines:
-                                                        st.code(line, language=None)
-
-                                                # Wait a moment so user can see the stats
-                                                import time
-                                                time.sleep(2)
-
-                                                st.info("🔄 Reloading with enriched content...")
-                                                st.rerun()
-                                            else:
-                                                st.error(f"❌ Enrichment failed: {result.get('error', 'Unknown error')}")
-                                                st.warning("💡 You can still save the original file")
-                                        finally:
-                                            # Clean up temp file
-                                            try:
-                                                os.unlink(temp_path)
-                                            except:
-                                                pass
-
-                        # Buttons row
-                        col1, col2 = st.columns(2)
-
-                        with col1:
-                            if st.button("✅ Use This File", type="primary", key="use_upload"):
-                                st.session_state.phrase_list = phrases
-                                st.session_state.qp_phrase_position = 0
-                                st.session_state.phrase_selector_widget = 0  # Sync widget state
-                                st.session_state.state_change_log.append(f"Upload file: Reset position to 0")
-                                st.session_state.quick_last_result = None
-                                st.session_state.material_source = f"Uploaded: {uploaded_file.name}"
-                                st.session_state.qp_materials_expanded = False  # Close expander
-                                st.success("✅ File loaded! Scroll down to practice section.")
-                                # NO rerun - let Streamlit update naturally to preserve tab
-
-                        with col2:
-                            # Save to server button (only for authenticated users)
-                            if st.session_state.get('authenticated'):
-                                if st.button("💾 Save to Server", key="save_upload"):
-                                    import remote_storage
-
-                                    # Get username from user dict (not 'anonymous' fallback - only authenticated users see this button)
-                                    user = st.session_state.get('user', {})
-                                    username = user.get('username', 'unknown')
-                                    # Use material_language (short code like 'pt', 'fr') not language (full name)
-                                    current_lang = st.session_state.get('material_language', 'fr')
-
-                                    # Use current content from session state (may be enriched)
-                                    upload_key = f"upload_{uploaded_file.name}_{uploaded_file.size}"
-                                    content_to_save = st.session_state.get(upload_key, content)
-
-                                    with st.spinner("Uploading to server..."):
-                                        # Show debug info
-                                        st.caption(f"📤 Uploading as user: {username}, language: {current_lang}")
-                                        st.caption(f"📁 Target: ~/miolingo.io/public_ftp/incoming/{username}/{current_lang}/")
-
-                                        result = remote_storage.save_user_material(
-                                            content=content_to_save,
-                                            filename=uploaded_file.name,
-                                            language=current_lang,
-                                            username=username
-                                        )
-
-                                    if result['success']:
-                                        st.success(f"✅ Saved to server: {result['path']}")
-                                        st.caption(f"📊 {result['verification']}")
-
-                                        # Show quota info
-                                        try:
-                                            quota = remote_storage.get_user_quota(username)
-                                            st.info(f"📦 Your storage: {quota['used_mb']}/{quota['quota_mb']} MB used")
-                                        except Exception as e:
-                                            st.caption(f"⚠️ Could not check quota: {str(e)}")
-
-                                        # Mark that this file has been saved (so preview shows saved version)
-                                        st.session_state[f"{upload_key}_saved"] = True
-                                        st.success("🔄 File saved! Preview now shows server version.")
-                                        st.rerun()
-                                    else:
-                                        st.error(f"❌ Upload failed: {result.get('error', 'Unknown error')}")
-                                        # Show full error details in expander for debugging
-                                        if result.get('error'):
-                                            with st.expander("🔍 Error Details"):
-                                                st.code(result['error'])
-                            else:
-                                st.caption("💡 Login to save files to server")
-
-                    except Exception as e:
-                        st.error(f"Error reading file: {e}")
-
-        # Show current material source
-        if 'phrase_list' in st.session_state and st.session_state.phrase_list:
-            material_source = st.session_state.get('material_source', 'Unknown source')
-            st.info(f"📚 **Current material:** {material_source}")
-
-            if st.button("🗑️ Clear Material"):
-                st.session_state.phrase_list = []
-                st.session_state.qp_phrase_position = 0
-                st.session_state.phrase_selector_widget = 0  # Sync widget state
-                st.session_state.state_change_log.append(f"Clear material: Reset position to 0")
-                st.session_state.quick_last_result = None
-                st.session_state.material_source = None
-                st.rerun()
-
-        # Determine practice mode
-        guided_mode = 'phrase_list' in st.session_state and st.session_state.phrase_list
-
-        # MODE 1: Guided List Practice
-        if guided_mode:
-            st.markdown("---")
-            st.subheader("📚 Guided Practice Mode")
-
-            # Use global app state (initialized once at startup)
-            # No per-tab initialization needed - qp_phrase_position persists
-
-            # Progress and navigation
-            total_phrases = len(st.session_state.phrase_list)
-            current_idx = st.session_state.qp_phrase_position
-
-            # Keep index in bounds (e.g., if phrase list changes)
-            if current_idx < 0:
-                current_idx = 0
-                st.session_state.qp_phrase_position = 0
-                st.session_state.state_change_log.append(f"Tab load: Bounded qp_phrase_position to 0 (was negative)")
-            elif current_idx >= total_phrases:
-                current_idx = total_phrases - 1 if total_phrases > 0 else 0
-                st.session_state.qp_phrase_position = current_idx
-                st.session_state.state_change_log.append(f"Tab load: Bounded qp_phrase_position to {current_idx} (was >= {total_phrases})")
-            current_phrase_obj = st.session_state.phrase_list[current_idx]
-            # Handle both dict and string formats for backward compatibility
-            if isinstance(current_phrase_obj, dict):
-                current_phrase = current_phrase_obj['text']
-                phrase_translation = current_phrase_obj.get('translation')
-                phrase_ipa = current_phrase_obj.get('ipa')
-            else:
-                current_phrase = current_phrase_obj
-                phrase_translation = None
-                phrase_ipa = None
-
-            # Track phrase changes to show feedback
-            if 'last_phrase_index' not in st.session_state:
-                st.session_state.last_phrase_index = current_idx
-
-            if st.session_state.last_phrase_index != current_idx:
-                st.success(f"✓ Moved to phrase #{current_idx + 1}")
-                st.session_state.last_phrase_index = current_idx
-
-            # Progress bar
-            progress = (current_idx + 1) / total_phrases
-            st.progress(progress, text=f"Phrase {current_idx + 1} of {total_phrases}")
-
-            # Navigation buttons
-            # Check if we're in edit mode
-            in_edit_mode = st.session_state.get('edit_mode', False)
-
-            col1, col2, col3, col4 = st.columns([1, 1, 2, 1])
-            with col1:
-                # Disable navigation in edit mode to avoid confusion
-                prev_disabled = (current_idx == 0) or in_edit_mode
-                if st.button("⬅️ Previous", disabled=prev_disabled, key="nav_prev",
-                           help="Navigation disabled in edit mode" if in_edit_mode else None):
-                    # Update app state
-                    st.session_state.qp_phrase_position -= 1
-                    # Sync widget state so dropdown stays in sync
-                    st.session_state.phrase_selector_widget = st.session_state.qp_phrase_position
-                    st.session_state.state_change_log.append(f"Prev button: qp_phrase_position → {st.session_state.qp_phrase_position}")
-                    # Keep result when navigating
-                    st.rerun()
-            with col2:
-                # Disable navigation in edit mode to avoid confusion
-                next_disabled = (current_idx >= total_phrases - 1) or in_edit_mode
-                if st.button("Next ➡️", disabled=next_disabled, key="nav_next",
-                           help="Navigation disabled in edit mode" if in_edit_mode else None):
-                    # Update app state
-                    st.session_state.qp_phrase_position += 1
-                    # Sync widget state so dropdown stays in sync
-                    st.session_state.phrase_selector_widget = st.session_state.qp_phrase_position
-                    st.session_state.state_change_log.append(f"Next button: qp_phrase_position → {st.session_state.qp_phrase_position}")
-                    # Keep result when navigating
-                    st.rerun()
-            with col3:
-                def format_phrase(i):
-                    phrase_obj = st.session_state.phrase_list[i]
-                    phrase_text = phrase_obj['text'] if isinstance(phrase_obj, dict) else phrase_obj
-                    preview = f"{i+1}. {phrase_text[:40]}{'...' if len(phrase_text) > 40 else ''}"
-                    return preview
-
-                # Callback to sync widget selection to app state
-                def on_phrase_select():
-                    """When user changes dropdown, update app state from widget state"""
-                    new_pos = st.session_state.phrase_selector_widget
-                    st.session_state.qp_phrase_position = new_pos
-                    st.session_state.state_change_log.append(f"Dropdown: qp_phrase_position → {new_pos} (user selected)")
-
-                # Two-key pattern: app state (qp_phrase_position) + widget state (phrase_selector_widget)
-                # Widget reads from app state via index=, writes to widget state via key=
-                # on_change callback syncs widget state back to app state
-                st.selectbox(
-                    "Jump to phrase:",
-                    options=range(total_phrases),
-                    index=st.session_state.qp_phrase_position,  # Read from app state
-                    format_func=format_phrase,
-                    key="phrase_selector_widget",  # Widget state (separate from app state)
-                    on_change=on_phrase_select,  # Sync back to app state
-                    disabled=in_edit_mode,
-                    help="Phrase navigation disabled in edit mode" if in_edit_mode else "Jump directly to any phrase"
-                )
-            with col4:
-                # Edit button - disabled when in edit mode
-                if 'edit_mode' not in st.session_state:
-                    st.session_state.edit_mode = False
-
-                # Disable Edit button when already in edit mode (grayed out)
-                if st.button("✏️ Edit", key="toggle_edit",
-                            help="Edit current phrase or type your own",
-                            disabled=st.session_state.edit_mode):
-                    st.session_state.edit_mode = True
-                    st.rerun()
-
-            # Diagnostic expander (collapsible, for debugging state management)
-            with st.expander("🔍 State Diagnostics (for debugging)", expanded=False):
-                st.markdown("""
-                **Purpose**: Verify state persistence across tab switches and widget interactions.
-
-                This shows how `qp_phrase_position` (app state) and `phrase_selector_widget` (widget state)
-                are managed separately but kept in sync via callbacks.
-                """)
-
-                col_diag1, col_diag2 = st.columns(2)
-                with col_diag1:
-                    st.write("**Current State:**")
-                    st.json({
-                        "qp_phrase_position (app)": st.session_state.qp_phrase_position,
-                        "phrase_selector_widget": st.session_state.get('phrase_selector_widget', 'Not created yet'),
-                        "active_tab": st.session_state.get('active_tab', 'Unknown'),
-                        "edit_mode": st.session_state.get('edit_mode', False),
-                        "total_phrases": len(st.session_state.phrase_list) if st.session_state.get('phrase_list') else 0
-                    })
-
-                with col_diag2:
-                    st.write("**State Change Log (last 10):**")
-                    recent_log = st.session_state.state_change_log[-10:] if st.session_state.state_change_log else ["(No changes yet)"]
-                    for entry in reversed(recent_log):
-                        st.text(entry)
-
-                if st.button("Clear Log", key="clear_state_log"):
-                    st.session_state.state_change_log = []
-                    st.rerun()
-
-            st.markdown("---")
-
-            # Display current phrase - editable or fixed
-            if st.session_state.edit_mode:
-                st.markdown("### ✏️ Edit Mode:")
-                st.caption("Edit the phrase below or type something completely different")
-                text = st.text_input(
-                    "Phrase to practice:",
-                    value=current_phrase,
-                    key="edit_phrase_input"
-                )
-                if st.button("📚 Return to Guided Mode", key="back_to_guided"):
-                    st.session_state.edit_mode = False
-                    st.rerun()
-            else:
-                # Show translation/IPA if available (above phrase for mobile visibility)
-                source_lang = st.session_state.source_language
-                target_lang = st.session_state.target_language
-
-                translation_text = None
-                if source_lang == "English" and phrase_translation:
-                    translation_text = phrase_translation
-                elif source_lang != target_lang:
-                    translation_text = get_translation_from_llm(current_phrase, target_lang, source_lang)
-
-                if translation_text or phrase_ipa:
-                    with st.expander("📖 Translation & Reference", expanded=False):
-                        if translation_text and not translation_text.startswith('[error'):
-                            st.markdown(f"**{source_lang}:** {translation_text}")
-                        if phrase_ipa:
-                            st.markdown(f"**📚 Reference IPA ({target_lang}):** {format_ipa(phrase_ipa)}", unsafe_allow_html=True)
-                            st.caption("Compare with eSpeak IPA generated below")
-
-                # Display phrase - mobile-friendly with emoji inline
-                st.markdown(f"#### 🎯 **{current_phrase}**")
-
-                # Use this phrase for practice
-                text = current_phrase
-
-        # MODE 2: Free Text Practice
-        else:
-            st.write("Practice any word or phrase you like")
-
-            # Show navigation buttons (disabled) for consistency
-            col1, col2, col3 = st.columns([1, 1, 4])
-            with col1:
-                st.button("⬅️ Previous", disabled=True, key="nav_prev_disabled",
-                         help="Navigation only available in guided mode")
-            with col2:
-                st.button("Next ➡️", disabled=True, key="nav_next_disabled",
-                         help="Navigation only available in guided mode")
-            with col3:
-                st.write("")  # Spacing
-
-            st.markdown("---")
-            text = st.text_input("Enter word or phrase:", key="practice_text_free")
-
-        # Translation-aware practice text
-        source_lang = st.session_state.source_language
-        target_lang = st.session_state.target_language
-        direction = st.session_state.translation_direction
-
-        translated_text = None
-        practice_text = text
-
-        if text:
-            if direction == 'source_to_target':
-                translated_text = get_translation_from_llm(text, source_lang, target_lang)
-                if translated_text and not translated_text.startswith('[error'):
-                    practice_text = translated_text
-            else:
-                translated_text = get_translation_from_llm(text, target_lang, source_lang)
-
-        # Show translation + IPA reference for free text
-        if text and translated_text and not translated_text.startswith('[error'):
-            with st.expander("📖 Translation & Reference", expanded=False):
-                if direction == 'source_to_target':
-                    st.markdown(f"**{source_lang}:** {text}")
-                    st.markdown(f"**{target_lang}:** {translated_text}")
-                else:
-                    st.markdown(f"**{target_lang}:** {text}")
-                    st.markdown(f"**{source_lang}:** {translated_text}")
-
-                ipa = get_ipa_from_espeak(practice_text, get_language_code(target_lang))
-                if ipa and not ipa.startswith('[error'):
-                    st.markdown(f"**📚 Reference IPA ({target_lang}):** {format_ipa(ipa)}", unsafe_allow_html=True)
-
-        # Use reusable practice interface with quick practice key prefix
-        render_practice_interface(practice_text, key_prefix="quick")
-
-        # Show last result using reusable component
-        if st.session_state.get('quick_last_result'):
-            render_practice_results(st.session_state.quick_last_result, key_prefix="quick")
-
-            # Optional: Hear eSpeak phoneme pronunciation (local development only)
-            if IS_LOCAL_DEV and st.session_state.quick_last_result and not st.session_state.quick_last_result["exact_match"]:
-                st.markdown("---")
-                st.subheader("Compare Phoneme Sounds (eSpeak)")
-                st.caption("🔧 Development feature - requires local audio device")
-                col1, col2 = st.columns(2)
-                with col1:
-                    if st.button("🔊 Correct Phonemes", key="phoneme_correct"):
-                        speak_text(st.session_state.quick_last_result['target'],
-                                 voice=st.session_state.settings['voice'],
-                                 speed=st.session_state.settings['speed'],
-                                 pitch=st.session_state.settings['pitch'])
-                with col2:
-                    if st.button("🔊 Your Phonemes", key="phoneme_yours"):
-                        speak_text(st.session_state.quick_last_result['recognized'],
-                                 voice=st.session_state.settings['voice'],
-                                 speed=st.session_state.settings['speed'],
-                                 pitch=st.session_state.settings['pitch'])
+        render_quick_practice_tab()
 
     # Tab 2: Story Reader
     elif selected_tab_index == 1:
